requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
Flask-Login==0.6.2
Flask-SocketIO==5.3.4
python-socketio==5.9.0
python-engineio==4.7.1
twilio==8.10.0
openai==0.27.8
requests==2.31.0
python-dotenv==1.0.0
Werkzeug==2.3.7
WTForms==3.0.1
email-validator==2.0.0
